refactor: rename to `collections.rs`
`BTreeMap` is not a queue.
